+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 # Unreleased
 
 - Bump `@aptos-labs/aptos-cli` to `1.0.2`
+- Fix the `Move` CLI command to correctly handle the success/error outputs
 
 # 1.30.0 (2024-10-21)